Revert "🎨 Palette: Modernize Portfolio with Dynamic Data and UX Enhancements"
