+++ .gitignore
@@ -29,7 +29,6 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 .pnpm-debug.log*
-dev_server.log
 
 # env files (can opt-in for committing if needed)
 .env*
