--- dev_server.log
@@ -1,24 +1,0 @@
-
-> portfolio-app@0.1.1 dev /app
-> next dev
-
- ⚠ turbopack.root should be absolute, using: /app
-   ▲ Next.js 16.0.10 (Turbopack)
-   - Local:         http://localhost:3000
-   - Network:       http://192.168.0.2:3000
-
- ✓ Starting...
- ✓ Ready in 922ms
- ○ Compiling / ...
- GET / 200 in 6.8s (compile: 6.3s, render: 539ms)
- GET /api/github?username=umar-fr 200 in 2.1s (compile: 675ms, render: 1447ms)
- GET / 200 in 132ms (compile: 8ms, render: 124ms)
- GET /api/github?username=umar-fr 200 in 1324ms (compile: 4ms, render: 1321ms)
- GET / 200 in 244ms (compile: 11ms, render: 233ms)
- GET /api/github?username=umar-fr 200 in 1222ms (compile: 3ms, render: 1219ms)
- GET / 200 in 239ms (compile: 18ms, render: 221ms)
- GET /api/github?username=umar-fr 200 in 1334ms (compile: 4ms, render: 1330ms)
- GET / 200 in 130ms (compile: 8ms, render: 122ms)
- GET /api/github?username=umar-fr 200 in 1159ms (compile: 5ms, render: 1154ms)
- GET / 200 in 117ms (compile: 5ms, render: 112ms)
- GET /api/github?username=umar-fr 200 in 1347ms (compile: 3ms, render: 1344ms)
